use 'long' over 'size_t'

+++ hash_test.cpp
@@ -7,8 +7,8 @@ int main(int argc, char *argv[]) {
 	hash<string> hash_func;
 	hash<string> another_hash_func;
 
-	size_t hash = hash_func(test_str);
-	size_t another_hash = another_hash_func(test_str);
+	long hash = hash_func(test_str);
+	long another_hash = another_hash_func(test_str);
 	cout << hash << endl;
 	cout << another_hash << endl;
